+++ lib/scripts.php
@@ -6,26 +6,16 @@
  *
  * @package moaregenesis
  * @author  Antonio
- * @since 	1.0.0
+ * @since   1.0.0
  * @license GPL-2.0+
  * @link    http://studiomoare.com/
- * @version 2.0.2
+ * @version 2.0.3
  */
 
 // Enqueue Scripts and Styles.
 add_action( 'wp_enqueue_scripts', 'mg_enqueue_scripts_styles' );
 function mg_enqueue_scripts_styles() {
 
-	$appearance = genesis_get_config( 'appearance' );
-
-	wp_enqueue_style(
-		genesis_get_theme_handle() . '-fonts', 
-		$appearance['fonts-url'], 
-		[],
-		genesis_get_theme_version(),
-		null
-	);
-	
 	wp_enqueue_style(
 		'mainstyles', 
 		get_stylesheet_directory_uri() . "/assets/stylesheets/main.css",