+++ patterns/footer-columns.php
@@ -1,0 +1,39 @@
+<?php
+/**
+ * Title: Footer columns
+ * Slug: moaregenesis/footer-columns
+ * Categories: moarepatterns
+ * Keywords: Footer columns
+ */
+?>
+<!-- wp:group {"align":"full","style":{"spacing":{"padding":{"top":"6rem","bottom":"6rem"}}},"backgroundColor":"black","className":"mg-pattern-footer-columns "} -->
+<div class="wp-block-group alignfull mg-pattern-footer-columns has-black-background-color has-background" style="padding-top:6rem;padding-bottom:6rem"><!-- wp:columns {"className":"mb-0 wrap"} -->
+<div class="wp-block-columns mb-0 wrap"><!-- wp:column -->
+<div class="wp-block-column"><!-- wp:heading {"textColor":"white","className":"font-sans-serif ","fontSize":"medium"} -->
+<h2 class="wp-block-heading font-sans-serif has-white-color has-text-color has-medium-font-size">Contacte</h2>
+<!-- /wp:heading -->
+
+<!-- wp:paragraph {"textColor":"white"} -->
+<p class="has-white-color has-text-color">nf@nf.com</p>
+<!-- /wp:paragraph --></div>
+<!-- /wp:column -->
+
+<!-- wp:column -->
+<div class="wp-block-column"><!-- wp:html -->
+<p><svg xmlns="http://www.w3.org/2000/svg" width="200" viewBox="0 0 32 32" class="aligncenter"><path fill="#ffffff" d="M31.708 25.708c-0-0-0-0-0-0l-9.708-9.708 9.708-9.708c0-0 0-0 0-0 0.105-0.105 0.18-0.227 0.229-0.357 0.133-0.356 0.057-0.771-0.229-1.057l-4.586-4.586c-0.286-0.286-0.702-0.361-1.057-0.229-0.13 0.048-0.252 0.124-0.357 0.228 0 0-0 0-0 0l-9.708 9.708-9.708-9.708c-0-0-0-0-0-0-0.105-0.104-0.227-0.18-0.357-0.228-0.356-0.133-0.771-0.057-1.057 0.229l-4.586 4.586c-0.286 0.286-0.361 0.702-0.229 1.057 0.049 0.13 0.124 0.252 0.229 0.357 0 0 0 0 0 0l9.708 9.708-9.708 9.708c-0 0-0 0-0 0-0.104 0.105-0.18 0.227-0.229 0.357-0.133 0.355-0.057 0.771 0.229 1.057l4.586 4.586c0.286 0.286 0.702 0.361 1.057 0.229 0.13-0.049 0.252-0.124 0.357-0.229 0-0 0-0 0-0l9.708-9.708 9.708 9.708c0 0 0 0 0 0 0.105 0.105 0.227 0.18 0.357 0.229 0.356 0.133 0.771 0.057 1.057-0.229l4.586-4.586c0.286-0.286 0.362-0.702 0.229-1.057-0.049-0.13-0.124-0.252-0.229-0.357z"></path></svg></p>
+<!-- /wp:html --></div>
+<!-- /wp:column -->
+
+<!-- wp:column -->
+<div class="wp-block-column"><!-- wp:heading {"textAlign":"right","textColor":"white","className":"font-sans-serif ","fontSize":"medium"} -->
+<h2 class="wp-block-heading has-text-align-right font-sans-serif has-white-color has-text-color has-medium-font-size">Segueix-nos</h2>
+<!-- /wp:heading -->
+
+<!-- wp:social-links {"iconColor":"white","iconColorValue":"#ffffff","openInNewTab":true,"className":"is-style-logos-only","layout":{"type":"flex","justifyContent":"right"}} -->
+<ul class="wp-block-social-links has-icon-color is-style-logos-only"><!-- wp:social-link {"url":"https://www.instagram.com","service":"instagram"} /-->
+
+<!-- wp:social-link {"url":"https://twitter.com","service":"twitter"} /--></ul>
+<!-- /wp:social-links --></div>
+<!-- /wp:column --></div>
+<!-- /wp:columns --></div>
+<!-- /wp:group -->