+++ patterns/lead-parragraph-center.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ * Title: Lead parragraph
+ * Slug: moaregenesis/lead-parragraph
+ * Categories: moarepatterns
+ * Keywords: Lead parragrapah
+ */
+?>
+<!-- wp:group {"style":{"spacing":{"padding":{"top":"1.5rem","bottom":"1.5rem"}}},"className":"mg-pattern-lead-paragraph "} -->
+<div class="wp-block-group mg-pattern-lead-paragraph" style="padding-top:1.5rem;padding-bottom:1.5rem"><!-- wp:paragraph {"align":"left","style":{"typography":{"lineHeight":"1.1"}},"fontSize":"x-large"} -->
+<p class="has-text-align-left has-x-large-font-size" style="line-height:1.1">Sam, ilitiunt, sitionem delisimin nihicia sum renes nobis dendia quos volorrore es dit quo te nullori berendae et porestiam cuptatesti tem fugitio eium ne volum quam ut moluptate sam aut mi, simusam fuga.</p>
+<!-- /wp:paragraph --></div>
+<!-- /wp:group -->