+++ patterns/newsletter-00.php
@@ -1,0 +1,21 @@
+<?php
+/**
+ * Title: Newsletter 00
+ * Slug: moaregenesis/newsletter-00
+ * Categories: moarepatterns
+ * Keywords: Newsletter 00
+ */
+?>
+<!-- wp:columns {"verticalAlignment":"center","align":"full","style":{"spacing":{"margin":{"bottom":"0px","top":"0px"},"padding":{"top":"3rem","bottom":"3rem"}}},"backgroundColor":"gray","className":"mb-0 mg-pattern-newsletter-00 "} -->
+<div class="wp-block-columns alignfull are-vertically-aligned-center mb-0 mg-pattern-newsletter-00 has-gray-background-color has-background" style="margin-top:0px;margin-bottom:0px;padding-top:3rem;padding-bottom:3rem"><!-- wp:column {"verticalAlignment":"center"} -->
+<div class="wp-block-column is-vertically-aligned-center"><!-- wp:paragraph {"align":"right","textColor":"white","className":"font-sans-serif my-0 ","fontSize":"x-large"} -->
+<p class="has-text-align-right font-sans-serif my-0 has-white-color has-text-color has-x-large-font-size"><strong><?php echo esc_html__( 'Newsletter', 'moaregenesis' ); ?></strong> <?php echo esc_html_x( 'Sign in!', 'sample content for newsletter', 'moaregenesis' ); ?></p>
+<!-- /wp:paragraph --></div>
+<!-- /wp:column -->
+
+<!-- wp:column {"verticalAlignment":"center"} -->
+<div class="wp-block-column is-vertically-aligned-center"><!-- wp:paragraph -->
+<p>Form</p>
+<!-- /wp:paragraph --></div>
+<!-- /wp:column --></div>
+<!-- /wp:columns -->