+++ patterns/sub-footer-00.php
@@ -1,0 +1,35 @@
+<?php
+/**
+ * Title: Sub Footer 00
+ * Slug: moaregenesis/sub-footer-00
+ * Categories: moarepatterns
+ * Keywords: Sub Footer 00
+ */
+?>
+<!-- wp:group {"align":"full","style":{"spacing":{"padding":{"top":"1.5rem","bottom":"1.5rem"}}},"backgroundColor":"black","className":"mg-pattern-sub-footer-00 "} -->
+<div class="wp-block-group alignfull mg-pattern-sub-footer-00 has-black-background-color has-background" style="padding-top:1.5rem;padding-bottom:1.5rem"><!-- wp:columns {"className":"mb-0 wrap"} -->
+<div class="wp-block-columns mb-0 wrap"><!-- wp:column {"verticalAlignment":"center"} -->
+<div class="wp-block-column is-vertically-aligned-center"><!-- wp:list {"style":{"elements":{"link":{"color":{"text":"var:preset|color|white"}}}},"textColor":"white","className":"font-sans-serif list-inline my-0 ","fontSize":"small"} -->
+<ul class="font-sans-serif list-inline my-0 has-white-color has-text-color has-link-color has-small-font-size"><!-- wp:list-item -->
+<li><a href="#">Política de privacitat</a></li>
+<!-- /wp:list-item -->
+
+<!-- wp:list-item -->
+<li><a href="#">Política de cookies</a></li>
+<!-- /wp:list-item --></ul>
+<!-- /wp:list --></div>
+<!-- /wp:column -->
+
+<!-- wp:column -->
+<div class="wp-block-column"><!-- wp:paragraph {"align":"center","style":{"elements":{"link":{"color":{"text":"var:preset|color|white"}}}},"textColor":"white","className":"my-0 font-sans-serif","fontSize":"small"} -->
+<p class="has-text-align-center my-0 font-sans-serif has-white-color has-text-color has-link-color has-small-font-size"><a href="https://creativecommons.org/licenses/by-nc-nd/4.0/">Creative Commons BY-NC-ND 4.0</a></p>
+<!-- /wp:paragraph --></div>
+<!-- /wp:column -->
+
+<!-- wp:column -->
+<div class="wp-block-column"><!-- wp:paragraph {"align":"right","style":{"elements":{"link":{"color":{"text":"var:preset|color|white"}}}},"textColor":"white","className":"my-0 font-sans-serif","fontSize":"small"} -->
+<p class="has-text-align-right my-0 font-sans-serif has-white-color has-text-color has-link-color has-small-font-size">Lorem ipsum lorem</p>
+<!-- /wp:paragraph --></div>
+<!-- /wp:column --></div>
+<!-- /wp:columns --></div>
+<!-- /wp:group -->