+++ patterns/title-full-height-image.php
@@ -1,0 +1,21 @@
+<?php
+/**
+ * Title: Title Full Height Image
+ * Slug: moaregenesis/title-full-height-image
+ * Categories: moarepatterns
+ * Keywords: Title Full Height Image
+ */
+?>
+<!-- wp:columns {"verticalAlignment":"center","align":"full","style":{"spacing":{"margin":{"top":"0px","bottom":"0px"},"padding":{"top":"0px","right":"0px","bottom":"0px","left":"0px"}}},"backgroundColor":"black","className":"mg-pattern-title-full-height-image"} -->
+<div class="wp-block-columns alignfull are-vertically-aligned-center mg-pattern-title-full-height-image has-black-background-color has-background" style="margin-top:0px;margin-bottom:0px;padding-top:0px;padding-right:0px;padding-bottom:0px;padding-left:0px"><!-- wp:column {"verticalAlignment":"center","width":"50vw"} -->
+<div class="wp-block-column is-vertically-aligned-center" style="flex-basis:50vw"><!-- wp:heading {"textAlign":"center","style":{"spacing":{"margin":{"right":"1.5rem","left":"1.5rem"}}},"textColor":"white","fontSize":"gigantic"} -->
+<h2 class="wp-block-heading has-text-align-center has-white-color has-text-color has-gigantic-font-size" style="margin-right:1.5rem;margin-left:1.5rem">Lorem ipsum</h2>
+<!-- /wp:heading --></div>
+<!-- /wp:column -->
+
+<!-- wp:column {"verticalAlignment":"center","width":"50vw","className":"img-full-height ","layout":{"inherit":false}} -->
+<div class="wp-block-column is-vertically-aligned-center img-full-height" style="flex-basis:50vw"><!-- wp:image {"sizeSlug":"large","linkDestination":"none","className":"mb-0 "} -->
+<figure class="wp-block-image size-large mb-0"><img src="https://studiomoare.com/static/banner-5adbd687b71cb1cb1eb46c26945166ea.jpg" alt=""/></figure>
+<!-- /wp:image --></div>
+<!-- /wp:column --></div>
+<!-- /wp:columns -->